fix delete posts

+++ pages/api/delete_post.ts
@@ -4,26 +4,34 @@ import prisma from '@/lib/prisma'
 
 export default async function handler(req, res) {
 	const session = await getServerSession(req, res, authOptions)
-
+	const postId = req.body.id
 	if (!session) {
 		res.status(401).json({ error: 'Not authenticated' })
 		return
 	}
+
+	//confirm postId belongs to user
+	const posts = await prisma.post.findMany({
+		where: {
+			userId: session?.user?.userId,
+			id: postId
+		}
+	})
+
+	if (posts.length === 0) {
+		res.status(401).json({ error: 'Not authorized' })
+		return
+	}
+
 	try {
-		let postId = req.body.id
+		await prisma.message.deleteMany({ where: { postId } })
+		await prisma.chatPaywalls.deleteMany({ where: { postId } })
 		const result = await prisma.post.deleteMany({
 			where: {
 				id: postId,
 				userId: session?.user?.userId
 			}
 		})
-
-		if (result.count !== 1) {
-			res.status(401).json({ error: 'Not authorized' })
-			return
-		}
-		await prisma.message.deleteMany({ where: { postId } })
-		await prisma.chatPaywalls.deleteMany({ where: { postId } })
 
 		res.json(result)
 	} catch (e) {
